proof of concept: phx-portal
This is a proof of concept commit implementing a new phx binding called
`phx-portal`. This binding allows to designate an element to be rendered
at another location in the DOM, designated by the ID that `phx-portal`
points to. This can be useful to render things like dialogs at the top
layer, outside any containers that might affect their rendering (e.g.
overflow: hidden).

Nowadays there is also the Popover API and native `<dialog>` elements,
so this might not be that important any more.

Because of this, I also don't feel like this is something we really want
to support. This commit only shows how it could be implemented.